EZEE-2805: "Admin" tab is not the last one in AdminUI menu
